# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: AI Fullstack Starter API
  version: 1.0.0
  description: |
    Versioned, language-neutral API contract shared by all clients
    (web, mobile) and implemented by every backend (FastAPI reference
    today; Node/Express or others later).

    Rules (docs/architecture/ADR-001..ADR-004):
    - This file is the sole source of truth for the API boundary.
      Changes are additive-only within v1; breaking changes require
      /api/v2 plus an ADR.
    - Nothing language-specific crosses the boundary: errors use the
      Error code registry, never exception class names or framework
      shapes.
    - v1 has no authentication; user_id is an anonymous client-supplied
      scope, not an identity.
servers:
  - url: http://127.0.0.1:8000
    description: Local development
...
paths:
...
  /api/v1/agent/sessions:
    post:
      operationId: createAgentSession
      tags: [agent]
      summary: Create a new agent session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentSessionRequest"
            examples:
              withTitle:
                value:
                  title: "Fix the login bug"
              minimal:
                value: {}
      responses:
        "201":
          description: Agent session created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentSessionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/InvalidRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/agent/sessions/{session_id}:
    get:
      operationId: getAgentSession
      tags: [agent]
      summary: Retrieve an agent session by ID
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Agent session snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentSessionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/InvalidRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/agent/sessions/{session_id}/tasks:
    post:
      operationId: submitAgentTask
      tags: [agent]
      summary: Submit a coding task to an agent session
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAgentTaskRequest"
            examples:
              codingTask:
                value:
                  prompt: "Refactor the login handler to extract validation into a helper."
      responses:
        "202":
          description: Task accepted; agent execution dispatched.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentTaskResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/InvalidRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/agent/sessions/{session_id}/events:
    get:
      operationId: listAgentEvents
      tags: [agent]
      summary: List ordered events for an agent session
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
        - name: task_id
          in: query
          required: false
          schema:
            type: string
          description: Filter events to a specific task.
        - name: after_sequence
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
          description: Return only events with sequence greater than this value.
      responses:
        "200":
          description: Ordered list of agent events.
          content:
            application/json:
              schema:
                type: object
                required: [events]
                properties:
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/AgentEventResponse"
              examples:
                assistantOutput:
                  value:
                    events:
                      - event_id: "evt_001"
                        session_id: "sess_abc"
                        task_id: "task_xyz"
                        sequence: 1
                        event_type: "assistant_output"
                        payload: { content: "I will start by reading the login handler." }
                        created_at: "2026-09-25T10:00:01Z"
                toolCall:
                  value:
                    events:
                      - event_id: "evt_002"
                        session_id: "sess_abc"
                        task_id: "task_xyz"
                        sequence: 2
                        event_type: "tool_call"
                        payload: { tool: "read_file", input: { path: "src/auth/login.py" } }
                        created_at: "2026-09-25T10:00:02Z"
                toolResult:
                  value:
                    events:
                      - event_id: "evt_003"
                        session_id: "sess_abc"
                        task_id: "task_xyz"
                        sequence: 3
                        event_type: "tool_result"
                        payload: { tool: "read_file", output: "def login(request):\n    ..." }
                        created_at: "2026-09-25T10:00:03Z"
                approvalRequested:
                  value:
                    events:
                      - event_id: "evt_004"
                        session_id: "sess_abc"
                        task_id: "task_xyz"
                        sequence: 4
                        event_type: "approval_requested"
                        payload: { approval_id: "appr_001", action_type: "write_file", action_fingerprint: "sha256:abc123" }
                        created_at: "2026-09-25T10:00:04Z"
                errorEvent:
                  value:
                    events:
                      - event_id: "evt_005"
                        session_id: "sess_abc"
                        task_id: "task_xyz"
                        sequence: 5
                        event_type: "error"
                        payload: { message: "Provider timeout after 30s", code: "provider_error" }
                        created_at: "2026-09-25T10:00:05Z"
                completedEvent:
                  value:
                    events:
                      - event_id: "evt_006"
                        session_id: "sess_abc"
                        task_id: "task_xyz"
                        sequence: 6
                        event_type: "completed"
                        payload: { result: "Refactoring applied successfully." }
                        created_at: "2026-09-25T10:00:06Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/InvalidRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/agent/sessions/{session_id}/events/stream:
    get:
      operationId: streamAgentEvents
      tags: [agent]
      summary: Stream agent events as SSE for an agent session
      description: |
        Returns a `text/event-stream` (SSE-compatible) response emitting
        `AgentSseEvent` frames. Each frame carries an `event` name matching
        the event_type and a JSON `data` payload of `AgentEventResponse`.
        The stream closes after a terminal `error` or `completed` event.
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
        - name: task_id
          in: query
          required: false
          schema:
            type: string
          description: Restrict the stream to a specific task.
        - name: after_sequence
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
          description: Replay events from after this sequence number.
      responses:
        "200":
          description: |
            SSE event stream. Each frame has:
            - event: one of assistant_output | tool_call | tool_result |
              approval_requested | approval_decided | error | completed | task_accepted
            - data: JSON-serialised AgentEventResponse
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  SSE event stream. Terminal event names are `error` and
                  `completed`; clients must close the stream after either.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/InvalidRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/agent/sessions/{session_id}/approvals:
    get:
      operationId: listAgentApprovals
      tags: [agent]
      summary: List pending approval requests for an agent session
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, approved, rejected, expired, consumed]
          description: Filter approvals by status (defaults to pending).
      responses:
        "200":
          description: Approval requests for the session.
          content:
            application/json:
              schema:
                type: object
                required: [approvals]
                properties:
                  approvals:
                    type: array
                    items:
                      $ref: "#/components/schemas/ApprovalRequestResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/InvalidRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/agent/sessions/{session_id}/approvals/{approval_id}:
    post:
      operationId: decideAgentApproval
      tags: [agent]
      summary: Submit an approval decision for a pending agent action
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
        - name: approval_id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionRequest"
            examples:
              approve:
                value:
                  decision: approved
              reject:
                value:
                  decision: rejected
                  rationale: "This change looks risky."
      responses:
        "200":
          description: Updated approval request after the decision.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRequestResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/InvalidRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/agent/sessions/{session_id}/state:
    get:
      operationId: getAgentState
      tags: [agent]
      summary: Retrieve consolidated state for an agent session
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Consolidated session state snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentStateResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/InvalidRequest"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  responses:
    InvalidRequest:
      description: Request failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            emptyPrompt:
              value:
                error:
                  code: invalid_request
                  message: "prompt must be a non-empty string"
    Unauthorized:
      description: Missing or invalid authentication credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: Unexpected server-side failure.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable, language-neutral error code registry.
          enum:
            - invalid_request
            - rate_limited
            - provider_error
            - provider_unavailable
            - internal_error
        message:
          type: string
          description: Human-readable explanation, safe to display.
        details:
          type: object
          description: Optional structured context; additive only.
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
...
    CreateAgentSessionRequest:
      type: object
      description: Request body for creating a new agent session.
      properties:
        title:
          type: [string, "null"]
          description: Optional human-readable title for the session.
        metadata:
          type: [object, "null"]
          description: |
            Optional arbitrary key-value metadata stored with the session.
          additionalProperties: true
    SubmitAgentTaskRequest:
      type: object
      required: [prompt]
      description: Request body for submitting a coding task to an agent session.
      properties:
        prompt:
          type: string
          minLength: 1
          description: The coding task description or instruction for the agent.
        metadata:
          type: [object, "null"]
          description: Optional arbitrary metadata for the task.
          additionalProperties: true
    ApprovalDecisionRequest:
      type: object
      required: [decision]
      description: Request body for deciding a pending agent approval request.
      properties:
        decision:
          type: string
          enum: [approved, rejected]
          description: Whether to approve or reject the pending agent action.
        rationale:
          type: [string, "null"]
          description: Optional human-readable explanation for the decision.
    AgentSessionResponse:
      type: object
      required: [session_id, status, created_at, updated_at]
      description: Snapshot of an agent session.
      properties:
        session_id:
          type: string
          description: Unique identifier for the session.
        status:
          type: string
          enum: [active, idle, completed, error]
          description: Current lifecycle status of the session.
        title:
          type: [string, "null"]
          description: Optional human-readable title.
        created_at:
          type: string
          format: date-time
          description: ISO 8601 timestamp when the session was created.
        updated_at:
          type: string
          format: date-time
          description: ISO 8601 timestamp of the last session update.
        metadata:
          type: [object, "null"]
          description: Arbitrary metadata supplied at session creation.
          additionalProperties: true
    AgentTaskResponse:
      type: object
      required: [task_id, session_id, status, prompt, created_at, updated_at]
      description: Snapshot of a task submitted to an agent session.
      properties:
        task_id:
          type: string
          description: Unique identifier for the task.
        session_id:
          type: string
          description: Session this task belongs to.
        status:
          type: string
          enum: [pending, running, succeeded, failed, cancelled, timed_out, blocked]
          description: Current lifecycle status of the task.
        prompt:
          type: string
          description: The coding task instruction submitted by the developer.
        result:
          type: [string, "null"]
          description: Final result text when the task has succeeded.
        created_at:
          type: string
          format: date-time
          description: ISO 8601 timestamp when the task was created.
        updated_at:
          type: string
          format: date-time
          description: ISO 8601 timestamp of the last task status update.
    AgentEventResponse:
      type: object
      required: [event_id, session_id, task_id, sequence, event_type, created_at]
      description: |
        A single ordered event in the agent execution timeline. The `payload`
        field is a structured JSON object whose shape depends on `event_type`;
        it is absent for events that carry no additional data.
      properties:
        event_id:
          type: string
          description: Unique identifier for the event.
        session_id:
          type: string
          description: Session this event belongs to.
        task_id:
          type: string
          description: Task this event was produced by.
        sequence:
          type: integer
          minimum: 0
          description: |
            Monotonically increasing integer per session. Clients may
            use `after_sequence` to resume an event stream from a known
            position.
        event_type:
          type: string
          enum:
            - assistant_output
            - tool_call
            - tool_result
            - approval_requested
            - approval_decided
            - error
            - completed
            - task_accepted
            - agent_started
          description: |
            Discriminant for the event payload:
            - `assistant_output`: LLM-generated text fragment or plan step.
            - `tool_call`: Agent requested a tool invocation.
            - `tool_result`: Tool returned a result.
            - `approval_requested`: Agent requires human approval before
              proceeding with a mutating or unsafe action.
            - `approval_decided`: Developer approved or rejected an action.
            - `error`: Terminal event; execution failed.
            - `completed`: Terminal event; task finished successfully.
            - `task_accepted`: Non-terminal; task was accepted by the worker.
        payload:
          type: [object, "null"]
          description: |
            Structured JSON payload whose shape is determined by
            `event_type`. May be absent for events with no extra data.
          additionalProperties: true
        created_at:
          type: string
          format: date-time
          description: ISO 8601 timestamp when the event was persisted.
...
    ApprovalRequestResponse:
      type: object
      required:
        - approval_id
        - session_id
        - task_id
        - action_type
        - action_fingerprint
        - status
        - expires_at
        - created_at
        - updated_at
      description: |
        A durable approval request created when the agent proposes a
        mutating or unsafe action. Approval grants are single-use and
        bound to the action fingerprint.
      properties:
        approval_id:
          type: string
          description: Unique identifier for the approval request.
        session_id:
          type: string
          description: Session this approval belongs to.
        task_id:
          type: string
          description: Task that triggered the approval request.
        action_type:
          type: string
          description: |
            Logical name for the action requiring approval
            (e.g. write_file, run_command).
        action_payload:
          type: [object, "null"]
          description: |
            Structured description of the proposed action. May be
            absent if the action carries no additional parameters.
          additionalProperties: true
        action_fingerprint:
          type: string
          description: |
            Deterministic hash of the exact action parameters. An
            approval grant authorises only the matching fingerprint.
        status:
          type: string
          enum: [pending, approved, rejected, expired, consumed]
          description: Current lifecycle status of the approval request.
        expires_at:
          type: string
          format: date-time
          description: ISO 8601 timestamp after which the approval expires.
        consumed_at:
          type: [string, "null"]
          format: date-time
          description: ISO 8601 timestamp when the approval was consumed.
        created_at:
          type: string
          format: date-time
          description: ISO 8601 timestamp when the approval request was created.
        updated_at:
          type: string
          format: date-time
          description: ISO 8601 timestamp of the last approval status change.
    AgentStateResponse:
      type: object
      required: [session_id, session, tasks, events, pending_approvals]
      description: |
        Consolidated snapshot of an agent session including the session
        metadata, all tasks, recent events, and pending approvals.
        Intended for clients that want a single-request state refresh.
      properties:
        session_id:
          type: string
          description: Session identifier mirrored from the session object.
        session:
          $ref: "#/components/schemas/AgentSessionResponse"
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/AgentTaskResponse"
          description: All tasks for the session, ordered by creation time.
        events:
          type: array
          items:
            $ref: "#/components/schemas/AgentEventResponse"
          description: Recent events for the session, ordered by sequence.
        pending_approvals:
          type: array
          items:
            $ref: "#/components/schemas/ApprovalRequestResponse"
          description: Pending approval requests awaiting a developer decision.
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Bearer token for hosted mode. Local development endpoints operate
        without authentication (ADR-004). When a hosted deployment enables
        bearer authentication, protected paths return 401 if the token is
        absent or invalid.